requirements.txt
openai~=0.27.8
streamlit~=1.25.0
requests~=2.31.0
geopy~=2.3.0
docstring-parser~=0.15
streamlit-js-eval~=0.1.5
google-api-python-client~=2.94.0
google-auth-oauthlib~=0.4.6
streamlit-authenticator~=0.2.2
peewee~=3.16.2
PyYAML~=6.0
pandas~=2.0.3
beautifulsoup4~=4.12.2
html5lib~=1.1